requirements.txt
# Trading Platform — $100K Digital Wealth
streamlit>=1.32.0
plotly>=5.20.0
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0

# Robinhood official crypto API auth (Ed25519 signing)
PyNaCl>=1.5.0

# Claude AI signals (Anthropic API)
anthropic>=0.25.0

# Browser automation (for other tools)
playwright

# Optional: faster HTTP for data fetches
httpx>=0.27.0
